Fixed a bug when races were not loaded well. Added white icons and made a test of buttons when the data is not provided yet

+++ ErgastAPP/ErgastAPP/Models/Results/Result.cs
@@ -17,7 +17,7 @@ namespace ErgastAPP.Models
         public string PositionText { get; set; }
 
         [JsonProperty("points")]
-        public int Points { get; set; }
+        public double Points { get; set; }
 
         [JsonProperty("Driver")]
         public Driver Driver { get; set; }
